Fix biber options in Makefile
It seems that there is a breaking change for biber 1.8 version that is
only compatible to Trusty distributions. This means that for the 1.8
version, only `--output_directory` is accepted. For biber versions
2.4 and up, i.e. those compatible to Xenial distribution, you have
`--output-directory`. For higher versions, this is a non-breaking
change. Changing the dash into an underscore is the path of least
resistance compared to deploying a Docker container that contains an
Ubuntu Trusty distribution.

Signed-by: Lester James V. Miranda <ljvmiranda@gmail.com>